src/lambda/requirements.txt
boto3==1.26.137
PyJWT==2.6.0
bcrypt==4.0.1
